+++ app/Models/Patient.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Models;
+
+use App\Models\User;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class Patient extends Model
+{
+    use SoftDeletes;
+
+    public static $createRules = [];
+    
+    public static $updateRules = [];
+    
+    protected $fillable = ['user_id'];
+
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
+
+    public function appointments()
+    {
+        return $this->hasMany(Appointment::class);
+    }
+}