+++ database/migrations/2019_11_11_050754_create_appointments_table.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateAppointmentsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('appointments', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('patient_id');
+            $table->string('type')->nullable();
+            $table->dateTime('appointment_date')->index();
+            $table->dateTime('next_appointment_date')->nullable()->index();
+            $table->timestamps();
+            $table->softDeletes();
+            $table->foreign('patient_id')->references('id')->on('patients');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('appointments');
+    }
+}